Modified UserCtrl & User Service (from Baptiste com)

+++ src/controller/UserController.ts
@@ -17,7 +17,12 @@ export class UserController {
 
     async one(request: Request, response: Response, next: NextFunction) {
         try {
-            return await this.userService.one(+request.params.id)
+            const user = await this.userService.one(+request.params.id)
+            if (user) { return user }
+            return {
+                success: `ko`,
+                message: `user not found`
+            }
         } catch (error) {
             console.log("🐼UserController ~ one ~ error:", error)
         }
@@ -25,11 +30,19 @@ export class UserController {
 
     async save(request: Request, response: Response, next: NextFunction) {
         try {
-            
-            // appel d'une fonction qui cherche un user selon son email 
-            //si on obiten un résultat on sort de la fonction
+//? appel d'une fonction qui cherche un user selon son email 
+            // Check if email is already used
+            const user = await this.userService.oneByMail(request.body)
+            // IF email not exist
+            if (!user) {
+                return await this.userService.create(request.body)
+            }
 
-            return await this.userService.create(request.body)
+//? si on obtiens un résultat on sort de la fonction
+            return {
+                success: `ko`,
+                message: `email ${user.email} already used`
+            }
         } catch (error) {
             console.log("🐼UserController ~ save ~ error:", error)
         }
@@ -38,14 +51,23 @@ export class UserController {
     async remove(request: Request, response: Response, next: NextFunction) {
         try {
             // get by id l'user 
-            // const user = await this.userService.one(id)
-
-            // si ! user je sors
-            return await this.userService.remove(+request.params.id)
+            const user = await this.userService.one(+request.params.id)
+            if (user) {
+                await this.userService.remove(+request.params.id)
+                return {
+                    success: `ko`,
+                    message: `user '${user.nickname}' was deleted`
+                }
+            }
+            return {
+                success: `ko`,
+                message: `user not found`
+            }
         } catch (error) {
             console.log("🐼UserController ~ remove ~ error:", error)
         }
     }
+
     async update(request: Request, response: Response, next: NextFunction) {
         try {
             const id = +request.params.id;
